change SdpOrigin::sessV type uint32 -> uint64

+++ core/AmSdp.h
@@ -80,7 +80,7 @@ struct SdpOrigin
 {
   string user;
   unsigned int sessId;
-  unsigned int sessV;
+  uint64_t sessV;
   SdpConnection conn;
 
   SdpOrigin() : user(), conn() {}
